// generated/graphql.d.ts
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
};

export enum _ModelMutationType {
  Created = "CREATED",
  Updated = "UPDATED",
  Deleted = "DELETED"
}

/** Meta information about the query. */
export type _QueryMeta = {
  __typename?: "_QueryMeta";
  count: Scalars["Int"];
};
// ...
export type Film = Node & {
  __typename?: "Film";
  characters?: Maybe<Array<Person>>;
  createdAt: Scalars["DateTime"];
  /** The name of the director of this film. */
  director?: Maybe<Scalars["String"]>;
  /** The episode number of this film. */
  episodeId: Scalars["Int"];
  id: Scalars["ID"];
  /** indicates if the record is published */
  isPublished: Scalars["Boolean"];
  /** The opening paragraphs at the beginning of this film. */
  openingCrawl?: Maybe<Scalars["String"]>;
  planets?: Maybe<Array<Planet>>;
  /** The names of the producers of this film. */
  producers?: Maybe<Array<Scalars["String"]>>;
  /**  The ISO 8601 date format of film release at original creator country. */
  releaseDate?: Maybe<Scalars["DateTime"]>;
  species?: Maybe<Array<Species>>;
  starships?: Maybe<Array<Starship>>;
  /** The title of this film */
  title: Scalars["String"];
  updatedAt: Scalars["DateTime"];
  vehicles?: Maybe<Array<Vehicle>>;
  /** Meta information about the query. */
  _charactersMeta: _QueryMeta;
  /** Meta information about the query. */
  _planetsMeta: _QueryMeta;
  /** Meta information about the query. */
  _speciesMeta: _QueryMeta;
  /** Meta information about the query. */
  _starshipsMeta: _QueryMeta;
  /** Meta information about the query. */
  _vehiclesMeta: _QueryMeta;
};
// ...
export type Node = {
  __typename?: "Node";
  /** The id of the object. */
  id: Scalars["ID"];
};

/** A Person is an individual person or character within the Star Wars universe */
export type Person = Node & {
  __typename?: "Person";
  /** The birth year of the person, using the in-universe standard of BBY or ABY -
   * Before the Battle of Yavin or After the Battle of Yavin. The Battle of Yavin
   * is a battle that occurs at the end of Star Wars episode IV: A New Hope.
   */
  birthYear?: Maybe<Scalars["String"]>;
  createdAt: Scalars["DateTime"];
  /** The eye color of this person. Will be "UNKNOWN" if not known or null if the person does not have an eye. */
  eyeColor?: Maybe<Array<Person_Eye_Color>>;
  films?: Maybe<Array<Film>>;
  /**  The gender of this person. Will be "UNKNOWN" if not known or null if the person does not have a gender. */
  gender?: Maybe<Person_Gender>;
  /** The hair color of this person. Will be "UNKNOWN" if not known or null if the person does not have hair. */
  hairColor?: Maybe<Array<Person_Hair_Color>>;
  /** The height of the person in centimeters. */
  height?: Maybe<Scalars["Int"]>;
  homeworld?: Maybe<Planet>;
  id: Scalars["ID"];
  /** indicates if the record is published */
  isPublished: Scalars["Boolean"];
  /** The mass of the person in kilograms. */
  mass?: Maybe<Scalars["Float"]>;
  /** The name of this person. */
  name: Scalars["String"];
  /** The skin color of this person. */
  skinColor?: Maybe<Array<Person_Skin_Color>>;
  species?: Maybe<Array<Species>>;
  starships?: Maybe<Array<Starship>>;
  updatedAt: Scalars["DateTime"];
  vehicles?: Maybe<Array<Vehicle>>;
  /** Meta information about the query. */
  _filmsMeta: _QueryMeta;
  /** Meta information about the query. */
  _speciesMeta: _QueryMeta;
  /** Meta information about the query. */
  _starshipsMeta: _QueryMeta;
  /** Meta information about the query. */
  _vehiclesMeta: _QueryMeta;
};
// ...
export enum Person_Eye_Color {
  Unknown = "UNKNOWN",
  Blue = "BLUE",
  Yellow = "YELLOW",
  Red = "RED",
  Brown = "BROWN",
  Bluegrey = "BLUEGREY",
  Black = "BLACK",
  Orange = "ORANGE",
  Hazel = "HAZEL",
  Pink = "PINK",
  Gold = "GOLD",
  Green = "GREEN",
  White = "WHITE",
  Dark = "DARK"
}

/**  The gender of this person. Will be "UNKNOWN" if not known or null if the person does not have a gender. */
export enum Person_Gender {
  Unknown = "UNKNOWN",
  Male = "MALE",
  Female = "FEMALE",
  Hermaphrodite = "HERMAPHRODITE"
}

/** The hair color of this person. Will be "UNKNOWN" if not known or null if the person does not have hair. */
export enum Person_Hair_Color {
  Auburn = "AUBURN",
  Black = "BLACK",
  Blonde = "BLONDE",
  Brown = "BROWN",
  Grey = "GREY",
  Unknown = "UNKNOWN",
  White = "WHITE"
}

/** The skin color of this person. */
export enum Person_Skin_Color {
  Unknown = "UNKNOWN",
  Fair = "FAIR",
  Gold = "GOLD",
  White = "WHITE",
  Light = "LIGHT",
  Green = "GREEN",
  Greentan = "GREENTAN",
  Pale = "PALE",
  Metal = "METAL",
  Dark = "DARK",
  Brownmottle = "BROWNMOTTLE",
  Brown = "BROWN",
  Grey = "GREY",
  Mottledgreen = "MOTTLEDGREEN",
  Orange = "ORANGE",
  Blue = "BLUE",
  Red = "RED",
  Yellow = "YELLOW",
  Tan = "TAN",
  Silver = "SILVER"
}
// ...
export type Planet = Node & {
  __typename?: "Planet";
  /** The climate of this planet. */
  climate?: Maybe<Array<Scalars["String"]>>;
  createdAt: Scalars["DateTime"];
  /** The diameter of this planet in kilometers. */
  diameter?: Maybe<Scalars["Int"]>;
  films?: Maybe<Array<Film>>;
  /** A number denoting the gravity of this planet, where "1" is normal or 1
   * standard G. "2" is twice or 2 standard Gs. "0.5" is half or 0.5 standard Gs.
   */
  gravity?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  /** indicates if the record is published */
  isPublished: Scalars["Boolean"];
  /** The name of the planet */
  name: Scalars["String"];
  /** The number of standard days it takes for this planet to complete a single orbit of its local star. */
  orbitalPeriod?: Maybe<Scalars["Int"]>;
  /** The average population of sentient beings inhabiting this planet. */
  population?: Maybe<Scalars["Float"]>;
  residents?: Maybe<Array<Person>>;
  /** The number of standard hours it takes for this planet to complete a single rotation on its axis. */
  rotationPeriod?: Maybe<Scalars["Int"]>;
  /** The percentage of the planet surface that is naturally occurring water or bodies of water. */
  surfaceWater?: Maybe<Scalars["Float"]>;
  /** The terrain of this planet. */
  terrain?: Maybe<Array<Scalars["String"]>>;
  updatedAt: Scalars["DateTime"];
  /** Meta information about the query. */
  _filmsMeta: _QueryMeta;
  /** Meta information about the query. */
  _residentsMeta: _QueryMeta;
};
// ...
export type Species = Node & {
  __typename?: "Species";
  /** The average height of this species in centimeters. */
  averageHeight?: Maybe<Scalars["Int"]>;
  /** The average lifespan of this species in years. */
  averageLifespan?: Maybe<Scalars["Int"]>;
  /** The classification of this species, such as "mammal" or "reptile". */
  classification?: Maybe<Scalars["String"]>;
  createdAt: Scalars["DateTime"];
  /** The designation of this species, such as "sentient". */
  designation?: Maybe<Scalars["String"]>;
  /** The eye colors for this species, "UNKNOWN" if not known, null if this species does not typically have eyes. */
  eyeColor?: Maybe<Array<Species_Eye_Color>>;
  films?: Maybe<Array<Film>>;
  /** The hair colors for this species, "UNKNOWN" if not known, null if this species does not typically have hairs. */
  hairColor?: Maybe<Array<Species_Hair_Color>>;
  id: Scalars["ID"];
  /** indicates if the record is published */
  isPublished: Scalars["Boolean"];
  /** The language commonly spoken by this species. */
  language?: Maybe<Scalars["String"]>;
  /** The name of this species. */
  name: Scalars["String"];
  people?: Maybe<Array<Person>>;
  /** The skin colors for this species, "UNKNOWN" if not known, null if this species does not typically have a skin. */
  skinColor?: Maybe<Array<Species_Skin_Color>>;
  updatedAt: Scalars["DateTime"];
  /** Meta information about the query. */
  _filmsMeta: _QueryMeta;
  /** Meta information about the query. */
  _peopleMeta: _QueryMeta;
};
// ...
export enum Species_Eye_Color {
  Amber = "AMBER",
  Black = "BLACK",
  Blue = "BLUE",
  Brown = "BROWN",
  Gold = "GOLD",
  Green = "GREEN",
  Grey = "GREY",
  Hazel = "HAZEL",
  Indigo = "INDIGO",
  Orange = "ORANGE",
  Pink = "PINK",
  Red = "RED",
  Silver = "SILVER",
  Unknown = "UNKNOWN",
  Yellow = "YELLOW",
  Golden = "GOLDEN"
}

/** The hair colors for this species, "UNKNOWN" if not known, null if this species does not typically have hairs. */
export enum Species_Hair_Color {
  Unknown = "UNKNOWN",
  Brown = "BROWN",
  White = "WHITE",
  Red = "RED",
  Black = "BLACK",
  Blonde = "BLONDE"
}
// ...
export type Starship = Node & {
  __typename?: "Starship";
  /** The maximum number of kilograms that this starship can transport. */
  cargoCapacity?: Maybe<Scalars["Float"]>;
  /** The class of this starship, such as "Starfighter" or "Deep Space Mobile Battlestation" */
  class?: Maybe<Scalars["String"]>;
  /** The maximum length of time that this starship can provide consumables for its entire crew without having to resupply. */
  consumables?: Maybe<Scalars["String"]>;
  /**  The cost of this starship new, in galactic credits. */
  costInCredits?: Maybe<Scalars["Float"]>;
  createdAt: Scalars["DateTime"];
  /** The number of personnel needed to run or pilot this starship. */
  crew?: Maybe<Scalars["Int"]>;
  films?: Maybe<Array<Film>>;
  /** The class of this starships hyperdrive. */
  hyperdriveRating?: Maybe<Scalars["Float"]>;
  id: Scalars["ID"];
  /** indicates if the record is published */
  isPublished: Scalars["Boolean"];
  /** The length of this starship in meters. */
  length?: Maybe<Scalars["Float"]>;
  /** The manufacturer of this starship. */
  manufacturer?: Maybe<Array<Scalars["String"]>>;
  /**  The maximum speed of this starship in the atmosphere. null if this starship is incapable of atmospheric flight. */
  maxAtmospheringSpeed?: Maybe<Scalars["Int"]>;
  /** The Maximum number of Megalights this starship can travel in a standard hour.
   * A "Megalight" is a standard unit of distance and has never been defined before
   * within the Star Wars universe.
   */
  mglt?: Maybe<Scalars["Int"]>;
  /** The name of this starship. The common name, such as "Death Star". */
  name: Scalars["String"];
  /**  The number of non-essential people this starship can transport. */
  passengers?: Maybe<Scalars["Int"]>;
  pilots?: Maybe<Array<Person>>;
  updatedAt: Scalars["DateTime"];
  /** Meta information about the query. */
  _filmsMeta: _QueryMeta;
  /** Meta information about the query. */
  _pilotsMeta: _QueryMeta;
};
// ...
export type Vehicle = Node & {
  __typename?: "Vehicle";
  /** The maximum number of kilograms that this vehicle can transport. */
  cargoCapacity?: Maybe<Scalars["Int"]>;
  /**  The class of this vehicle, such as "Wheeled" or "Repulsorcraft". */
  class?: Maybe<Scalars["String"]>;
  /** The maximum length of time that this vehicle can provide consumables for its entire crew without having to resupply. */
  consumables?: Maybe<Scalars["String"]>;
  /** The cost of this vehicle new, in Galactic Credits. */
  costInCredits?: Maybe<Scalars["Int"]>;
  createdAt: Scalars["DateTime"];
  /** The number of personnel needed to run or pilot this vehicle. */
  crew?: Maybe<Scalars["Int"]>;
  films?: Maybe<Array<Film>>;
  id: Scalars["ID"];
  /** indicates if the record is published */
  isPublished: Scalars["Boolean"];
  /**  The length of this vehicle in meters. */
  length?: Maybe<Scalars["Float"]>;
  /** The manufacturer of this vehicle. */
  manufacturer?: Maybe<Array<Scalars["String"]>>;
  /** The maximum speed of this vehicle in the atmosphere. */
  maxAtmospheringSpeed?: Maybe<Scalars["Int"]>;
  /** The model or official name of this vehicle. Such as "All-Terrain Attack Transport". */
  model?: Maybe<Scalars["String"]>;
  /** The name of this vehicle. The common name, such as "Sand Crawler" or "Speeder bike". */
  name: Scalars["String"];
  /** The number of non-essential people this vehicle can transport. */
  passengers?: Maybe<Scalars["Int"]>;
  pilots?: Maybe<Array<Person>>;
  updatedAt: Scalars["DateTime"];
  /** Meta information about the query. */
  _filmsMeta: _QueryMeta;
  /** Meta information about the query. */
  _pilotsMeta: _QueryMeta;
};